docs: Split Scrapling guide example into modules and use code tabs
